Improve vite-like babel transforms for jest config
