+++ jest/babel-preset-vite/babel-preset-vite.js
@@ -1,0 +1,8 @@
+module.exports = function () {
+  return {
+    plugins: [
+      require("./babel-plugin-syntax-vite-meta-glob.js"),
+      require("./babel-plugin-syntax-vite-meta-env.js"),
+    ],
+  };
+};
