+++ jest/babel-preset-vite/index.js
@@ -1,0 +1,1 @@
+module.exports = require("./babel-preset-vite");
