+++ jest/babelTransform.js
@@ -5,16 +5,16 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
+/* eslint-disable @typescript-eslint/no-var-requires */
 "use strict";
 
 const babelJest = require("babel-jest");
-const importMetaBabelPlugin = require("./importMetaBabelPlugin");
 
 module.exports = babelJest.createTransformer({
   presets: [
     "babel-preset-react-app",
     "@babel/preset-react",
     "@babel/preset-typescript",
+    require("./babel-preset-vite"),
   ],
-  plugins: [[importMetaBabelPlugin]],
 });
